JAMES-2812 Improve Guice structure for DeletedMessageVault
 - Do not leak Cassandra dependency everywhere
 - Rework module split for the vault
