+++ server/container/guice/mailbox-plugin-deleted-messages-vault-guice/pom.xml
@@ -39,10 +39,6 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
-            <artifactId>apache-james-mailbox-deleted-messages-vault-cassandra</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-deleted-messages-vault</artifactId>
         </dependency>
         <dependency>
