+++ server/container/guice/mailbox-plugin-deleted-messages-vault-guice/src/main/java/org/apache/james/modules/vault/DeletedMessageVaultModule.java
@@ -19,8 +19,6 @@
 
 package org.apache.james.modules.vault;
 
-import java.io.FileNotFoundException;
-
 import javax.inject.Singleton;
 
 import org.apache.commons.configuration.Configuration;
@@ -31,7 +29,6 @@ import org.apache.james.mailrepository.memory.MailRepositoryStoreConfiguration;
 import org.apache.james.utils.PropertiesProvider;
 import org.apache.james.vault.DeletedMessageVault;
 import org.apache.james.vault.MailRepositoryDeletedMessageVault;
-import org.apache.james.vault.RetentionConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,6 +43,8 @@ public class DeletedMessageVaultModule extends AbstractModule {
 
     @Override
     protected void configure() {
+        install(new DeletedMessageVaultRetentionModule());
+
         bind(MailRepositoryDeletedMessageVault.class).in(Scopes.SINGLETON);
         bind(DeletedMessageVault.class)
             .to(MailRepositoryDeletedMessageVault.class);
@@ -67,15 +66,4 @@ public class DeletedMessageVaultModule extends AbstractModule {
         }
     }
 
-    @Provides
-    @Singleton
-    RetentionConfiguration providesRetentionConfiguration(PropertiesProvider propertiesProvider) throws ConfigurationException, org.apache.commons.configuration.ConfigurationException {
-        try {
-            Configuration configuration = propertiesProvider.getConfiguration("deletedMessageVault");
-            return RetentionConfiguration.from(configuration);
-        } catch (FileNotFoundException e) {
-            LOGGER.warn("Error encountered while retrieving Deleted message vault configuration. Using default MailRepository RetentionTime (1 year) instead.");
-            return RetentionConfiguration.DEFAULT;
-        }
-    }
 }
